✨ Improved API use and cost + Documentation
Now we are giving some memory / buffer to the DeepSeek API. The API is stateless so we need some memory. And so we can get a better response aware of a bigger context (10k tokens) and with a max price per prompt of 2x the max price of the first prompt. Gives better results with an optimal price